Update README.md Lists to include The Herd Locker
Based on the hundreds of thousands of #slack tweets already processed, noise is a little over 87% of the conversation. The platform filters and ranks the most popular shared content in realtime. Slack's zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, tracking dozens of topics. It will always be free to use. No ads. No BS.